+++ doc/whatsnew.rst
@@ -6,6 +6,7 @@ Next release
 
 - :func:`.bare.create_res` accepts a caller-supplied spec and data (:pull:`559`).
 - Support the 2025 edition of the Climate Policy Database in :mod:`.tools.newclimate` (:pull:`560`).
+- New class :class:`.SDMXSource` for exogenous data from SDMX web services, using :func:`.util.sdmx.fetch_data`; new modules :ref:`tools-eurostat` and :ref:`tools-unsd` for openly-licensed energy balance data (:pull:`543`).
 - Fix import of :mod:`message_ix_models` without :mod:`plotnine` (:issue:`323`, :issue:`540`, :pull:`547`).
 - Add IAMC code list :class:`~.iamc.structure.CL_SCENARIO_DIAGNOSTIC` (:pull:`501`).
 - New module :ref:`tools-newclimate` (:pull:`499`).